Restructured project code into packages (by type) for better organization.
