--- Android/app/src/main/java/com/spybug/gtnav/State.java
@@ -1,8 +1,0 @@
-package com.spybug.gtnav;
-
-public enum State {
-    MAIN,
-    DIRECTIONS,
-    BUSES,
-    BIKES
-}
