+++ Android/app/src/main/java/com/spybug/gtnav/fragments/BikesOverlayFragment.java
@@ -1,4 +1,4 @@
-package com.spybug.gtnav;
+package com.spybug.gtnav.fragments;
 
 import android.content.Context;
 import android.net.Uri;
@@ -8,6 +8,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageButton;
+import android.widget.Switch;
+import android.widget.Toast;
+
+import com.spybug.gtnav.models.BikeStation;
+import com.spybug.gtnav.utils.BikeStationsServerRequest;
+import com.spybug.gtnav.interfaces.Communicator;
+import com.spybug.gtnav.interfaces.OnEventListener;
+import com.spybug.gtnav.R;
+
+import java.util.List;
 
 
 /**
@@ -18,11 +28,12 @@ import android.widget.ImageButton;
  * Use the {@link MainMapOverlayFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class MainMapOverlayFragment extends Fragment {
+public class BikesOverlayFragment extends Fragment {
     private OnFragmentInteractionListener mListener;
     private ImageButton menuButton;
+    private View view;
 
-    public MainMapOverlayFragment() {
+    public BikesOverlayFragment() {
         // Required empty public constructor
     }
 
@@ -35,8 +46,8 @@ public class MainMapOverlayFragment extends Fragment {
      * @return A new instance of fragment DirectionsFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static MainMapOverlayFragment newInstance(String param1, String param2) {
-        MainMapOverlayFragment fragment = new MainMapOverlayFragment();
+    public static BikesOverlayFragment newInstance(String param1, String param2) {
+        BikesOverlayFragment fragment = new BikesOverlayFragment();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
@@ -50,18 +61,58 @@ public class MainMapOverlayFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.fragment_main_map_overlay, container, false);
-        
-        menuButton = v.findViewById(R.id.search_bar_menubutton);
-        menuButton.setOnClickListener(new View.OnClickListener() {
+        view = inflater.inflate(R.layout.fragment_bikes_overlay, container, false);
+        final Switch stationSwitch = view.findViewById(R.id.bike_station_switch);
+        stationSwitch.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                ((MainActivity) getActivity()).openDrawer(); //opens the navigation drawer
+                if (!stationSwitch.isChecked()) {
+                    hideBikeStations();
+                    stationSwitch.setText(getString(R.string.show_station_switch_text));
+                } else {
+                    showBikeStations();
+                    stationSwitch.setText(getString(R.string.show_racks_switch_text));
+                }
             }
         });
 
-        return v;
+        getBikeStations();
+
+
+        return view;
     }
+
+    public void getBikeStations() {
+        new BikeStationsServerRequest(getContext(), new OnEventListener<List<BikeStation>, String>() {
+            @Override
+            public void onSuccess(List<BikeStation> bikeStations) {
+                Communicator communicator = (Communicator) getActivity();
+                if (communicator != null) {
+                    communicator.passBikeStationsToMap(bikeStations);
+                }
+            }
+
+            @Override
+            public void onFailure(String message) {
+                Toast.makeText(view.getContext(), message, Toast.LENGTH_LONG).show();
+            }
+        }).execute();
+    }
+
+    /**
+     * Hides the bike stations on the map and displays the bike rack locations
+     */
+    private void hideBikeStations() {
+        //TODO
+    }
+
+    /**
+     * Hides the bike rack locations on the map and shows the bike stations
+     */
+    private void showBikeStations() {
+        //TODO
+    }
+
 
     @Override
     public void onAttach(Context context) {
