+++ Android/app/src/main/java/com/spybug/gtnav/interfaces/Communicator.java
@@ -1,6 +1,9 @@
-package com.spybug.gtnav;
+package com.spybug.gtnav.interfaces;
 
 import com.mapbox.mapboxsdk.geometry.LatLng;
+import com.spybug.gtnav.models.BikeStation;
+import com.spybug.gtnav.models.Bus;
+
 import java.util.List;
 
 public interface Communicator {
