+++ Android/app/src/main/java/com/spybug/gtnav/utils/BikeStationsServerRequest.java
@@ -1,11 +1,11 @@
-package com.spybug.gtnav;
+package com.spybug.gtnav.utils;
 
 import android.content.Context;
 import android.os.AsyncTask;
 
-import com.mapbox.mapboxsdk.geometry.LatLng;
-import com.mapbox.services.commons.models.Position;
-import com.mapbox.services.commons.utils.PolylineUtils;
+import com.spybug.gtnav.BuildConfig;
+import com.spybug.gtnav.interfaces.OnEventListener;
+import com.spybug.gtnav.models.BikeStation;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -20,24 +20,24 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.spybug.gtnav.HelperUtil.haveNetworkConnection;
+import static com.spybug.gtnav.utils.HelperUtil.haveNetworkConnection;
 
 
 /**
- * Background task to get a bus route
+ * Background task to communicate with the map server
  */
 
-public class BusRouteServerRequest extends AsyncTask<Object, Void, List<List<LatLng>>> {
+public class BikeStationsServerRequest extends AsyncTask<Object, Void, List<BikeStation>> {
 
-    private WeakReference<Context> contextRef;
     private static final String REQUEST_METHOD = "GET";
     private static final int READ_TIMEOUT = 15000;
     private static final int CONNECTION_TIMEOUT = 15000;
+    private WeakReference<Context> contextRef;
     private boolean hasNetwork = true;
     private int errorCode = 0;
-    private OnEventListener<List<List<LatLng>>, String> mCallBack;
+    private OnEventListener<List<BikeStation>, String> mCallBack;
 
-    BusRouteServerRequest(Context context, OnEventListener<List<List<LatLng>>, String> callback) {
+    public BikeStationsServerRequest(Context context, OnEventListener<List<BikeStation>, String> callback) {
         contextRef = new WeakReference<>(context);
         mCallBack = callback;
     }
@@ -47,41 +47,40 @@ public class BusRouteServerRequest extends AsyncTask<Object, Void, List<List<Lat
     }
 
     @Override
-    protected List<List<LatLng>> doInBackground(Object[] objects) {
-        List<List<LatLng>> points = new ArrayList<>();
-        String routeTag = (String)objects[0];
+    protected List<BikeStation> doInBackground(Object[] objects) {
+        List<BikeStation> bikeStations = new ArrayList<>();
 
         if (!hasNetwork) {
             errorCode = 1;
-            return points;
+            return bikeStations;
         }
 
         String inputLine;
         String stringUrl;
         String result;
 
-        stringUrl = String.format("%sroutes?route=%s",
-                BuildConfig.API_URL,
-                routeTag);
+        stringUrl = String.format("%sbikes", BuildConfig.API_URL);
 
         try {
-
             URL myUrl = new URL(stringUrl);
-            HttpURLConnection connection = (HttpURLConnection) myUrl.openConnection();
+            HttpURLConnection connection =(HttpURLConnection) myUrl.openConnection();
 
             connection.setRequestMethod(REQUEST_METHOD);
             connection.setReadTimeout(READ_TIMEOUT);
             connection.setConnectTimeout(CONNECTION_TIMEOUT);
-            connection.connect();
 
+            connection.connect();
             InputStreamReader streamReader = new InputStreamReader(connection.getInputStream());
+
             BufferedReader reader = new BufferedReader(streamReader);
             StringBuilder stringBuilder = new StringBuilder();
             while((inputLine = reader.readLine()) != null){
                 stringBuilder.append(inputLine);
             }
+
             reader.close();
             streamReader.close();
+
             result = stringBuilder.toString();
         }
         catch(IOException e){
@@ -91,28 +90,27 @@ public class BusRouteServerRequest extends AsyncTask<Object, Void, List<List<Lat
 
         if (result != null) {
             try {
-                JSONObject resultJSON = new JSONObject(result);
-                JSONArray encodedStrings = resultJSON.getJSONArray("route");
-                for (int i = 0; i < encodedStrings.length(); i++) {
-                    List<Position> positionList = PolylineUtils.decode(encodedStrings.getString(i), 5);
-                    List<LatLng> newList = new ArrayList<LatLng>();
+                JSONArray stations = new JSONArray(result);
 
-                    // Convert Positions List into List<LatLng>
-                    for (int j = 0; j < positionList.size(); j++) {
-                        newList.add(new LatLng(
-                                positionList.get(j).getLatitude(),
-                                positionList.get(j).getLongitude()));
-                    }
-                    points.add(newList);
+                for (int i = 0; i < stations.length(); i++) {
+                    JSONObject station = stations.getJSONObject(i);
+                    BikeStation newStation = new BikeStation(station.getString("station_id"),
+                            station.getString("name"),
+                            station.getDouble("lat"), station.getDouble("lon"),
+                            station.getInt("num_bikes_available"), station.getInt("num_bikes_disabled"),
+                            station.getInt("num_docks_available"));
+                    bikeStations.add(newStation);
                 }
-            } catch (JSONException e) {
-                e.printStackTrace();
+            }
+            catch(JSONException ex) {
+                ex.printStackTrace();
             }
         }
-        return points;
+
+        return bikeStations;
     }
 
-    protected void onPostExecute(List<List<LatLng>> result) {
+    protected void onPostExecute(List<BikeStation> result) {
         if (mCallBack != null) {
             if (errorCode != 0) {
                 if (errorCode == 1) {
