+++ Android/app/src/main/res/layout/fragment_bikes_overlay.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.spybug.gtnav.MainMapOverlayFragment">
+    tools:context="com.spybug.gtnav.fragments.MainMapOverlayFragment">
 
 
     <Switch
