+++ Android/app/src/main/res/layout/fragment_bottom_navbar.xml
@@ -4,7 +4,7 @@
              android:layout_height="wrap_content"
              android:gravity="center"
              android:layout_gravity="bottom"
-             tools:context="com.spybug.gtnav.BottomNavbarFragment">
+             tools:context="com.spybug.gtnav.fragments.BottomNavbarFragment">
 
     <LinearLayout
         android:id="@+id/bottom_bar"
