+++ Android/app/src/main/res/layout/fragment_schedule.xml
@@ -2,7 +2,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context="com.spybug.gtnav.ScheduleFragment">
+             tools:context="com.spybug.gtnav.fragments.ScheduleFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
